requirements.txt
matplotlib
pandas
Shapely
descartes
imageio
Pillow
natsort